Create Exception , API

+++ src/main/java/org/example/spring_jpa/service/impl/UserServiceImpl.java
@@ -1,6 +1,7 @@
 package org.example.spring_jpa.service.impl;
 
 import org.example.spring_jpa.dto.UserDto;
+import org.example.spring_jpa.exception.DuplicateFieldException;
 import org.example.spring_jpa.exception.NotFoundException;
 import org.example.spring_jpa.model.User;
 import org.example.spring_jpa.repository.UserRepository;
@@ -11,8 +12,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
+import org.springframework.web.bind.MethodArgumentNotValidException;
 
 import java.time.LocalDateTime;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 @Service
@@ -26,12 +31,14 @@ public class UserServiceImpl implements UserService {
         this.userRepository = userRepository;
         this.modelMapper = modelMapper;
     }
+
     @Override
     public Iterable<User> getAllUsers()
     {
         return userRepository.findAll(Sort.by(Sort.Direction.ASC,"lastname"));
 
     }
+
     @Override
     public User getUserById(Integer id)
     {
@@ -63,9 +70,30 @@ public class UserServiceImpl implements UserService {
     @Override
     public void saveUser(User user)
     {
-         userRepository.save(user);
+        userRepository.save(user);
     }
 
+    @Override
+    public void saveUserDto(UserDto userDto)
+    {
+        Map<String, String> errors = new HashMap<>();
+
+        if (userRepository.existsUserByEmail(userDto.getEmail())) {
+            errors.put("email", "Email already exists");
+        }
+
+        if (userRepository.existsUserByPhone(userDto.getPhone())) {
+            errors.put("phone", "Phone already exists");
+        }
+
+        if (!errors.isEmpty()) {
+            throw new DuplicateFieldException(errors);
+        }
+        User user = modelMapper.map(userDto, User.class);
+        if(user.getCreatedDate()==null)
+            user.setCreatedDate(LocalDateTime.now());
+        userRepository.save(user);
+    }
     @Override
     public void deleteUser(User user)
     {
@@ -75,6 +103,9 @@ public class UserServiceImpl implements UserService {
     @Override
     public void deleteUserById(int id)
     {
+        if(!userRepository.existsById(id)) {
+            throw new NotFoundException("User not found");
+        }
         userRepository.deleteById(id);
     }
 
@@ -84,5 +115,39 @@ public class UserServiceImpl implements UserService {
         User user = modelMapper.map(userDto, User.class);
         user.setCreatedDate(LocalDateTime.now());
         userRepository.save(user);
+    }
+
+    @Override
+    public User createUser(UserDto userDto)
+    {
+        User user = modelMapper.map(userDto, User.class);
+        user.setCreatedDate(LocalDateTime.now());
+        return userRepository.save(user);
+    }
+
+    @Override
+    public Optional<User> updateUser(Integer id , UserDto userDto)
+    {
+        Map<String, String> errors = new HashMap<>();
+
+        if(userRepository.existsUserByEmail(userDto.getEmail())) {
+            errors.put("email", "Email already exists");
+        }
+        if(userRepository.existsUserByPhone(userDto.getPhone())) {
+            errors.put("phone", "Phone already exists");
+        }
+        if (!errors.isEmpty()) {
+            throw new DuplicateFieldException(errors);
+        }
+        return userRepository.findById(id)
+                .map(existingUser -> {
+                    existingUser.setLastname(userDto.getLastname());
+                    existingUser.setFirstname(userDto.getFirstname());
+                    existingUser.setUsername(userDto.getUsername());
+                    existingUser.setPhone(userDto.getPhone());
+                    existingUser.setEmail(userDto.getEmail());
+                    existingUser.setPassword(userDto.getPassword());
+                    return userRepository.save(existingUser);
+                });
     }
 }
